requirements.txt
pyotp
docopt
qrcode
requests
pycryptodome
